Flyttar beregnings-knytta kode over i beregningsservicen
