+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/hentinformasjon/beregning/BeregningService.kt
@@ -5,8 +5,9 @@ import no.nav.etterlatte.brev.behandling.Avkortingsinfo
 import no.nav.etterlatte.brev.behandling.Beregningsperiode
 import no.nav.etterlatte.brev.behandling.Utbetalingsinfo
 import no.nav.etterlatte.brev.behandling.hentUtbetaltBeloep
-import no.nav.etterlatte.brev.hentinformasjon.hentBenyttetTrygdetidOgProratabroek
+import no.nav.etterlatte.libs.common.IntBroek
 import no.nav.etterlatte.libs.common.behandling.SakType
+import no.nav.etterlatte.libs.common.beregning.BeregningsMetode
 import no.nav.etterlatte.libs.common.feilhaandtering.UgyldigForespoerselException
 import no.nav.etterlatte.libs.common.vedtak.VedtakType
 import no.nav.etterlatte.libs.ktor.token.BrukerTokenInfo
@@ -144,3 +145,44 @@ class BeregningService(
         )
     }
 }
+
+private fun hentBenyttetTrygdetidOgProratabroek(
+    beregningsperiode: no.nav.etterlatte.libs.common.beregning.Beregningsperiode,
+): Pair<Int, IntBroek?> =
+    when (beregningsperiode.beregningsMetode) {
+        BeregningsMetode.NASJONAL ->
+            Pair(
+                beregningsperiode.samletNorskTrygdetid ?: throw SamletTeoretiskTrygdetidMangler(),
+                null,
+            )
+
+        BeregningsMetode.PRORATA -> {
+            Pair(
+                beregningsperiode.samletTeoretiskTrygdetid ?: throw SamletTeoretiskTrygdetidMangler(),
+                beregningsperiode.broek ?: throw BeregningsperiodeBroekMangler(),
+            )
+        }
+
+        BeregningsMetode.BEST -> throw UgyldigBeregningsMetode()
+        null -> beregningsperiode.trygdetid to null
+    }
+
+class SamletTeoretiskTrygdetidMangler :
+    UgyldigForespoerselException(
+        code = "SAMLET_TEORETISK_TRYGDETID_MANGLER",
+        detail = "Samlet teoretisk trygdetid mangler i beregningen",
+    )
+
+class BeregningsperiodeBroekMangler :
+    UgyldigForespoerselException(
+        code = "BEREGNINGSPERIODE_BROEK_MANGLER",
+        detail = "Beregningsperioden mangler brøk",
+    )
+
+class UgyldigBeregningsMetode :
+    UgyldigForespoerselException(
+        code = "UGYLDIG_BEREGNINGS_METODE",
+        detail =
+            "Kan ikke ha brukt beregningsmetode 'BEST' i en faktisk beregning, " +
+                "siden best velger mellom nasjonal eller prorata når det beregnes.",
+    )
